--- anonymous-feedback/.browserslistrc
@@ -1,13 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-env",
-      "react-app",
-      {
-        "targets": {
-            "esmodules": true
-        }
-      }
-    ]
-  ]
-}
