+++ anonymous-feedback/package.json
@@ -3,7 +3,7 @@
   "private": true,
   "license": "MIT",
   "dependencies": {
-    "@babel/preset-env": "7.10.4",
+    "@babel/preset-env": "^7.11.5",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
@@ -50,6 +50,9 @@
   "author": "aidan parkinson",
   "devDependencies": {
     "@babel/core": "^7.11.6",
-    "@babel/register": "^7.11.5"
+    "@babel/plugin-transform-react-jsx": "^7.10.4",
+    "@babel/preset-react": "^7.10.4",
+    "@babel/register": "^7.11.5",
+    "babel-preset-es2015": "^6.24.1"
   }
 }
