+++ anonymous-feedback/server/index.js
@@ -1,7 +1,12 @@
-const express = require('express');
+var React = require('react');
+var ReactDOMServer = require('react-dom/server');
 
-// we'll talk about this in a minute:
-const serverRenderer = require('./middleware/renderer');
+// import our main App component
+var App = require('../src/App');
+
+var fs = require('fs');
+
+var express = require('express');
 
 const PORT = 3000;
 const path = require('path');
@@ -11,7 +16,29 @@ const app = express();
 const router = express.Router();
 
 // root (/) should always serve our server rendered page
-router.use('^/$', serverRenderer);
+router.use('^/$', function (req, res, next) {
+
+    // point to the html file created by CRA's build tool
+    const filePath = path.resolve(__dirname, '..', '..', 'build', 'index.html');
+
+    fs.readFile(filePath, 'utf8', (err, htmlData) => {
+        if (err) {
+            console.error('err', err);
+            return res.status(404).end()
+        }
+
+        // render the app as a string
+        const html = ReactDOMServer.renderToString(<App />);
+
+        // inject the rendered app into our html and send it
+        return res.send(
+            htmlData.replace(
+                '<div id="root"></div>',
+                `<div id="root">${html}</div>`
+            )
+        );
+    });
+});
 
 // other static resources should just be served as they are
 router.use(express.static(
